Reference

Types

MLFlowClient.MLFlow — Type

MLFlow

Base type which defines location and version for MLFlow API service.

Fields

baseuri::String: base MLFlow tracking URI, e.g. http://localhost:5000

apiversion: used API version, e.g. 2.0

Constructors

MLFlow(baseuri; apiversion=2.0)

Examples

julia> mlf = MLFlow("http://localhost:5000")
MLFlow("http://localhost:5000", 2.0)

MLFlowClient.MLFlowExperiment — Type

MLFlowExperiment

Represents an MLFlow experiment.

Fields

name::String: experiment name.

lifecycle_stage::String: life cycle stage, one of ["active", "deleted"]

experiment_id::Integer: experiment identifier.

tags::Any: list of tags.

artifact_location::String: where are experiment artifacts stored.

Constructors

MLFlowExperiment(name, lifecycle_stage, experiment_id, tags, artifact_location)

MLFlowExperiment(exp::Dict{String,Any})

MLFlowClient.MLFlowRun — Type

MLFlowRun

Represents an MLFlow run.

Fields

info::MLFlowRunInfo: Run metadata.

data::MLFlowRunData: Run data.

Constructors

MLFlowRun(rundata::MLFlowRunData)

MLFlowRun(runinfo::MLFlowRunInfo)

MLFlowRun(info::Dict{String,Any})

MLFlowRun(info::Dict{String,Any}, data::Dict{String,Any})

MLFlowClient.MLFlowRunInfo — Type

MLFlowRunInfo

Represents run metadata.

Fields

run_id::String: run identifier.

experiment_id::Integer: experiment identifier.

status::MLFlowRunStatus: run status.

start_time::Union{Int64,Missing}: when was the run started, UNIX time in milliseconds.

end_time::Union{Int64,Missing}: when did the run end, UNIX time in milliseconds.

artifact_uri::String: where are artifacts from this run stored.

lifecycle_stage::String: one of active or deleted.

Constructors

MLFlowRunInfo(run_id, experiment_id, status, start_time, end_time, artifact_uri, lifecycle_stage)

MLFlowRunInfo(info::Dict{String,Any})

MLFlowClient.MLFlowRunData — Type

MLFlowRunData

Represents run data.

Fields

metrics::Vector{MLFlowRunDataMetric}: run metrics.

params::Dict{String,String}: run parameters.

tags: list of run tags.

Constructors

MLFlowRunData(data::Dict{String,Any})

MLFlowClient.MLFlowRunDataMetric — Type

MLFlowRunDataMetric

Represents a metric.

Fields

key::String: metric identifier.

value::Float64: metric value.

step::Int64: step.

timestamp::Int64: timestamp in UNIX time in milliseconds.

Constructors

MLFlowRunDataMetric(d::Dict{String,Any})

MLFlowClient.MLFlowRunStatus — Type

MLFlowRunStatus

Represents the status of an MLFlow Run.

Fields

status::String: one of RUNNING/SCHEDULED/FINISHED/FAILED/KILLED

Constructors

MLFlowRunStatus(status::String)

Experiments

MLFlowClient.createexperiment — Function

createexperiment(mlf::MLFlow; name=missing, artifact_location=missing, tags=missing)

Creates an MLFlow experiment.

Arguments

mlf: MLFlow configuration.

name: experiment name. If not specified, MLFlow sets it.

artifact_location: directory where artifacts of this experiment will be stored. If not specified, MLFlow uses its default configuration.

tags: a Dictionary of key-values which tag the experiment.

Returns

Experiment identifier (integer).

MLFlowClient.getexperiment — Function

getexperiment(mlf::MLFlow, experiment_id::Integer)

Retrieves an MLFlow experiment by id.

Arguments

mlf: MLFlow configuration.

experiment_id: Experiment identifier.

Returns

An instance of type MLFlowExperiment

getexperiment(mlf::MLFlow, experiment_name::String)

Retrieves an MLFlow experiment by name.

Arguments

mlf: MLFlow configuration.

experiment_name: Experiment name.

Returns

An instance of type MLFlowExperiment

MLFlowClient.listexperiments — Function

listexperiments(mlf::MLFlow)

Returns a list of MLFlow experiments.

TODO: not yet entirely implemented

MLFlowClient.deleteexperiment — Function

deleteexperiment(mlf::MLFlow, experiment_id)

Deletes an MLFlow experiment.

Arguments

mlf: MLFlow configuration.

experiment_id: experiment identifier.

Runs

MLFlowClient.createrun — Function

createrun(mlf::MLFlow, experiment_id; start_time=missing, tags=missing)

Creates a run associated to an experiment.

Arguments

mlf: MLFlow configuration.

experiment_id: experiment identifier.

Keywords

start_time: if provided, must be a UNIX timestamp in milliseconds. By default, set to current time.

tags: if provided, must be a key-value structure such as a dictionary.

Returns

an instance of type MLFlowRun

MLFlowClient.getrun — Function

getrun(mlf::MLFlow, run_id)

Retrieves information about an MLFlow run.

Arguments

mlf: MLFlow configuration.

run_id::String: run identifier.

Returns

an instance of type MLFlowRun

MLFlowClient.updaterun — Function

updaterun(mlf::MLFlow, run, status; end_time=missing)

Updates the status of an experiment's run.

Arguments

mlf: MLFlow configuration.

run: one of MLFlowRun, MLFlowRunInfo, or String.

status: either String and one of ["RUNNING", "SCHEDULED", "FINISHED", "FAILED", "KILLED"], or an instance of MLFlowRunStatus

Keywords

end_time: if provided, must be a UNIX timestamp in milliseconds. By default, set to current time.

MLFlowClient.deleterun — Function

deleterun(mlf::MLFlow, run)

Deletes an experiment's run.

Arguments

mlf: MLFlow configuration.

run: one of MLFlowRun, MLFlowRunInfo, or String.

MLFlowClient.searchruns — Function

searchruns(mlf::MLFlow, experiment_ids)

Searches for runs in an experiment.

Arguments

mlf: MLFlow configuration.

experiment_ids::AbstractVector{Integer}: experiment_ids in which to search for runs. Can also be a single Integer.

Keywords

filter::String: filter as defined in MLFlow documentation

run_view_type::String: one of ACTIVE_ONLY, DELETED_ONLY, or ALL.

max_results::Integer: 50,000 by default.

order_by::String: as defined in MLFlow documentation

page_token::String: paging functionality, handled automatically. Not meant to be passed by the user.

Returns

vector of MLFlowRun runs that were found in the list of experiments.

Logging

MLFlowClient.logparam — Function

logparam(mlf::MLFlow, run, key, value)
logparam(mlf::MLFlow, run, kv)

Associates a key/value pair of parameters to the particular run.

Arguments

mlf: MLFlow configuration.

run: one of MLFlowRun, MLFlowRunInfo, or String.

key: parameter key (name).

value: parameter value.

One could also specify kv::Dict instead of separate key and value arguments.

MLFlowClient.logmetric — Function

logmetric(mlf::MLFlow, run, key, value::T; timestamp, step) where T<:Real
logmetric(mlf::MLFlow, run, key, values::AbstractArray{T}; timestamp, step) where T<:Real

Logs a metric value (or values) against a particular run.

Arguments

mlf: MLFlow configuration.

run: one of MLFlowRun, MLFlowRunInfo, or String

key: metric name.

value: metric value, must be numeric.

Keywords

timestamp: if provided, must be a UNIX timestamp in milliseconds. By default, set to current time.

step: step at which the metric value has been taken.

MLFlowClient.logartifact — Function

logartifact(mlf::MLFlow, run, filename)

Stores an artifact (file) in the run's artifact location.

Note

Assumes that artifact_uri is mapped to a local directory. At the moment, this only works if both MLFlow and the client are running on the same host or they map a directory that leads to the same location over NFS, for example.

Arguments

mlf::MLFlow: MLFlow onfiguration. Currently not used, but when this method is extended to support S3, information from mlf will be needed.

run: one of MLFlowRun, MLFlowRunInfo or String.

filename: path to the artifact that needs to be sent to MLFlow.