Experiment operations

MLFlowClient.createexperiment — Function

createexperiment(instance::MLFlow, name::String;
    artifact_location::Union{String, Missing}=missing,
    tags::MLFlowUpsertData{Tag}=Tag[])

Create an Experiment with a name. Returns the newly created Experiment. Validates that another Experiment with the same name does not already exist and fails if another Experiment with the same name already exists.

Arguments

instance: MLFlow configuration.

name: Experiment name. This field is required.

artifact_location: Location where all artifacts for the Experiment are stored. If not provided, the remote server will select an appropriate default.

tags: A collection of Tag to set on the Experiment.

Returns

The ID of the newly created Experiment.

MLFlowClient.getexperiment — Function

getexperiment(instance::MLFlow, experiment_id::String)
getexperiment(instance::MLFlow, experiment_id::Integer)

Get metadata for an Experiment. This method works on deleted experiments.

Arguments

instance: MLFlow configuration.

experiment_id: ID of the associated Experiment.

Returns

An instance of type Experiment.

MLFlowClient.getexperimentbyname — Function

getexperimentbyname(instance::MLFlow, experiment_name::String)

Get metadata for an Experiment.

This endpoint will return deleted experiments, but prefers the active Experiment if an active and deleted Experiment share the same name. If multiple deleted experiments share the same name, the API will return one of them.

Arguments

instance: MLFlow configuration.

experiment_name: Name of the associated Experiment.

Returns

An instance of type Experiment.

MLFlowClient.deleteexperiment — Function

deleteexperiment(instance::MLFlow, experiment_id::String)
deleteexperiment(instance::MLFlow, experiment_id::Integer)
deleteexperiment(instance::MLFlow, experiment::Experiment)

Mark an Experiment and associated metadata, runs, metrics, params, and tags for deletion. If the Experiment uses FileStore, artifacts associated with Experiment are also deleted.

Arguments

instance: MLFlow configuration.

experiment_id: ID of the associated Experiment.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.restoreexperiment — Function

restoreexperiment(instance::MLFlow, experiment_id::String)
restoreexperiment(instance::MLFlow, experiment_id::Integer)
restoreexperiment(instance::MLFlow, experiment::Experiment)

Restore an Experiment marked for deletion. This also restores associated metadata, runs, metrics, params, and tags. If Experiment uses FileStore, underlying artifacts associated with Experiment are also restored.

Arguments

instance: MLFlow configuration.

experiment_id: ID of the associated Experiment.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.updateexperiment — Function

updateexperiment(instance::MLFlow, experiment_id::String, new_name::String)
updateexperiment(instance::MLFlow, experiment_id::Integer, new_name::String)
updateexperiment(instance::MLFlow, experiment::Experiment, new_name::String)

Update Experiment metadata.

Arguments

instance: MLFlow configuration.

experiment_id: ID of the associated Experiment.

new_name: If provided, the Experiment name is changed to the new name. The new name must be unique.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.searchexperiments — Function

searchexperiments(instance::MLFlow; max_results::Int64=20000, page_token::String="",
    filter::String="", order_by::Array{String}=String[],
    view_type::ViewType=ACTIVE_ONLY)

Arguments

instance: MLFlow configuration.

max_results: Maximum number of experiments desired.

page_token: Token indicating the page of experiments to fetch.

filter: A filter expression over Experiment attributes and tags that allows returning a subset of experiments. See MLFlow documentation.

order_by: List of columns for ordering search results, which can include Experiment name and id with an optional “DESC” or “ASC” annotation, where “ASC” is the default.

view_type: Qualifier for type of experiments to be returned. If unspecified, return only active experiments. For more values, see ViewType.

Returns

Vector of Experiment that were found in the MLFlow instance.

The next page token if there are more results.

MLFlowClient.setexperimenttag — Function

setexperimenttag(instance::MLFlow, experiment_id::String, key::String, value::String)
setexperimenttag(instance::MLFlow, experiment_id::Integer, key::String, value::String)
setexperimenttag(instance::MLFlow, experiment::Experiment, key::String, value::String)

Set a tag on an Experiment. Experiment tags are metadata that can be updated.

Arguments

experiment_id: ID of the Experiment under which to log the tag.

key: Name of the tag.

value: String value of the tag being logged.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.createexperimentpermission — Function

createexperimentpermission(instance::MLFlow, experiment_id::String, username::String,
    permission::Permission)
createexperimentpermission(instance::MLFlow, experiment_id::Integer, username::String,
    permission::Permission)
createexperimentpermission(instance::MLFlow, experiment::Experiment, username::String,
    permission::Permission)

Arguments

instance: MLFlow configuration.

experiment_id: Experiment id.

username: User username.

permission: Permission to grant.

Returns

An instance of type ExperimentPermission.

MLFlowClient.getexperimentpermission — Function

getexperimentpermission(instance::MLFlow, experiment_id::String, username::String)
getexperimentpermission(instance::MLFlow, experiment_id::Integer, username::String)
getexperimentpermission(instance::MLFlow, experiment::Experiment, username::String)

Arguments

instance: MLFlow configuration.

experiment_id: Experiment id.

username: User username.

Returns

An instance of type ExperimentPermission.

MLFlowClient.updateexperimentpermission — Function

updateexperimentpermission(instance::MLFlow, experiment_id::String, username::String,
    permission::Permission)
updateexperimentpermission(instance::MLFlow, experiment_id::Integer, username::String,
    permission::Permission)
updateexperimentpermission(instance::MLFlow, experiment::Experiment, username::String,
    permission::Permission)

Arguments

instance: MLFlow configuration.

experiment_id: Experiment id.

username: User username.

permission: Permission to grant.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.deleteexperimentpermission — Function

deleteexperimentpermission(instance::MLFlow, experiment_id::String, username::String)
deleteexperimentpermission(instance::MLFlow, experiment_id::Integer, username::String)
deleteexperimentpermission(instance::MLFlow, experiment::Experiment, username::String)

Arguments

instance: MLFlow configuration.

experiment_id: Experiment id.

username: User username.

Returns

true if successful. Otherwise, raises exception.