Miscellaneous operations

MLFlowClient.getmetrichistory — Function

getmetrichistory(instance::MLFlow, run_id::String, metric_key::String;
    page_token::String="", max_results::Union{Int64, Missing}=missing)

Get a list of all values for the specified Metric for a given Run.

Arguments

instance: MLFlow configuration.

run_id: ID of the Run from which to fetch Metric values.

metric_key: Name of the Metric to fetch.

page_token: Token indicating the page of Metric history to fetch.

max_results: Maximum number of logged instances of a Metric for a Run to return per call.

Returns

A list of all historical values for the specified Metric in the specified Run.

The next page token if there are more results.

MLFlowClient.refresh — Function

refresh(instance::MLFlow, run::Run)
refresh(instance::MLFlow, experiment::Experiment)

Get the latest metadata for a Run or Experiment.

Arguments

instance: MLFlow configuration.

run or experiment: Run or Experiment to refresh.

Returns

An instance of type Run or Experiment.