Registered model operations

MLFlowClient.createregisteredmodel — Function

createregisteredmodel(instance::MLFlow, name::String;
    tags::MLFlowUpsertData{Tag}=Tag[], description::Union{String, Missing}=missing)

Create a RegisteredModel with a name. Returns the newly created RegisteredModel. Validates that another RegisteredModel with the same name does not already exist and fails if another RegisteredModel with the same name already exists.

Arguments

instance: MLFlow configuration.

name: Register models under this name.

tags: A collection of Tag.

description: Optional description for RegisteredModel.

Returns

An instance of type RegisteredModel.

MLFlowClient.getregisteredmodel — Function

getregisteredmodel(instance::MLFlow, name::String)

Arguments

instance: MLFlow configuration.

name: RegisteredModel model unique name identifier.

Returns

An instance of type RegisteredModel.

MLFlowClient.renameregisteredmodel — Function

renameregisteredmodel(instance::MLFlow, name::String, new_name::String)

Arguments

instance: MLFlow configuration.

name: RegisteredModel unique name identifier.

new_name: If provided, updates the name for this RegisteredModel.

Returns

An instance of type RegisteredModel.

MLFlowClient.updateregisteredmodel — Function

updateregisteredmodel(instance::MLFlow, name::String;
    description::Union{String, Missing}=missing)

Arguments

instance: MLFlow configuration.

name: RegisteredModel unique name identifier.

description: If provided, updates the description for this RegisteredModel.

Returns

An instance of type RegisteredModel.

MLFlowClient.deleteregisteredmodel — Function

deleteregisteredmodel(instance::MLFlow, name::String)

Arguments

instance: MLFlow configuration.

name: RegisteredModel unique name identifier.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.searchregisteredmodels — Function

searchregisteredmodels(instance::MLFlow, filter::String, max_results::Int64,
    order_by::String, page_token::String)

Arguments

instance: MLFlow configuration.

filter: String filter condition. See MLFlow documentation.

max_results: Maximum number of models desired.

order_by: List of columns for ordering search results, which can include model name and last updated timestamp with an optional “DESC” or “ASC” annotation, where “ASC” is the default. Tiebreaks are done by model name ASC.

page_token: Pagination token to go to the next page based on a previous search query.

Returns

Vector of RegisteredModel that were found in the MLFlow instance.

The next page token if there are more results.

MLFlowClient.setregisteredmodeltag — Function

setregisteredmodeltag(instance::MLFlow, name::String, key::String, value::String)

Arguments

instance: MLFlow configuration.

name: Unique name of the model.

key: Name of the Tag.

value: String value of the Tag being logged.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.deleteregisteredmodeltag — Function

deleteregisteredmodeltag(instance::MLFlow, name::String, key::String)

Arguments

instance: MLFlow configuration.

name: Name of the RegisteredModel that the tag was logged under.

key: Name of the Tag.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.deleteregisteredmodelalias — Function

deleteregisteredmodelalias(instance::MLFlow, name::String, alias::String)

Arguments

instance: MLFlow configuration.

name: Name of the RegisteredModel.

alias: Name of the alias.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.setregisteredmodelalias — Function

setregisteredmodelalias(instance::MLFlow, name::String, alias::String, version::String)

Arguments

instance: MLFlow configuration.

name: Name of the RegisteredModel.

alias: Name of the alias.

version: ModelVersion number.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.createregisteredmodelpermission — Function

createregisteredmodelpermission(instance::MLFlow, name::String, username::String,
    permission::Permission)

Arguments

instance: MLFlow configuration.

name: RegisteredModel name.

username: User username.

permission: Permission to grant.

Returns

An instance of type RegisteredModelPermission.

MLFlowClient.getregisteredmodelpermission — Function

getregisteredmodelpermission(instance::MLFlow, name::String, username::String)

Arguments

instance: MLFlow configuration.

name: RegisteredModel name.

username: User username.

Returns

An instance of type RegisteredModelPermission.

MLFlowClient.updateregisteredmodelpermission — Function

updateregisteredmodelpermission(instance::MLFlow, name::String, username::String,
    permission::Permission)

Arguments

instance: MLFlow configuration.

name: RegisteredModel name.

username: User username.

permission: New permission to grant.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.deleteregisteredmodelpermission — Function

deleteregisteredmodelpermission(instance::MLFlow, name::String, username::String)

Arguments

instance: MLFlow configuration.

name: RegisteredModel name.

username: User username.

Returns

true if successful. Otherwise, raises exception.