Run operations

MLFlowClient.createrun — Function

createrun(instance::MLFlow, experiment_id::String;
    run_name::Union{String, Missing}=missing,
    start_time::Union{Int64, Missing}=missing,
    tags::Union{Dict{<:Any}, Array{<:Any}}=[])

Create a new Run within an Experiment. A Run is usually a single execution of a machine learning or data ETL pipeline.

Arguments

instance: MLFlow configuration.

experiment_id: ID of the associated Experiment.

run_name: Name of the Run.

start_time: Unix timestamp in milliseconds of when the Run started.

tags: Additional metadata for Run.

Returns

An instance of type Run.

MLFlowClient.deleterun — Function

deleterun(instance::MLFlow, run_id::String)
deleterun(instance::MLFlow, run::Run)

Mark a Run for deletion.

Arguments

instance: MLFlow configuration.

run_id: ID of the Run to delete.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.restorerun — Function

restorerun(instance::MLFlow, run_id::String)
restorerun(instance::MLFlow, run::Run)

Restore a deleted Run.

Arguments

instance: MLFlow configuration.

run_id: ID of the Run to restore.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.getrun — Function

getrun(instance::MLFlow, run_id::String)

Get metadata, metrics, params, and tags for a Run. In the case where multiple metrics with the same key are logged for a Run, return only the value with the latest timestamp. If there are multiple values with the latest timestamp, return the maximum of these values.

Arguments

instance: MLFlow configuration.

run_id: ID of the Run to fetch.

Returns

An instance of type Run.

MLFlowClient.setruntag — Function

setruntag(instance::MLFlow, run_id::String, key::String, value::String)
setruntag(instance::MLFlow, run::Run, key::String, value::String)
setruntag(instance::MLFlow, run::Run, tag::Tag)

Set a Tag on a Run.

Arguments

instance: MLFlow configuration.

run_id: ID of the Run under which to log the Tag.

key: Name of the Tag.

value: String value of the Tag being logged.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.deleteruntag — Function

deleteruntag(instance::MLFlow, run_id::String, key::String)
deleteruntag(instance::MLFlow, run::Run, key::String)
deleteruntag(instance::MLFlow, run::Run, tag::Tag)

Delete a Tag on a Run.

Arguments

instance: MLFlow configuration.

run_id: ID of the Run that the Tag was logged under.

key: Name of the Tag.

Returns

true if successful. Otherwise, raises exception.

MLFlowClient.searchruns — Function

searchruns(instance::MLFlow; experiment_ids::Array{String}=String[], filter::String="",
    run_view_type::ViewType=ACTIVE_ONLY, max_results::Int=1000,
    order_by::Array{String}=String[], page_token::String="")

Search for runs that satisfy expressions. Search expressions can use Metric and Param keys.

Arguments

instance: MLFlow configuration.

experiment_ids: List of Experiment IDs to search over.

filter: A filter expression over params, metrics, and tags, that allows returning a subset of runs. See MLFlow documentation.

run_view_type: Whether to display only active, only deleted, or all runs. Defaults to only active runs.

max_results: Maximum number of runs desired.

order_by: List of columns to be ordered by, including attributes, params, metrics, and tags with an optional “DESC” or “ASC” annotation, where “ASC” is the default.

page_token: Token indicating the page of runs to fetch.

Returns

Vector of Run that were found in the specified experiments.

The next page token if there are more results.

MLFlowClient.updaterun — Function

updaterun(instance::MLFlow, run_id::String; status::Union{RunStatus, Missing}=missing,
    end_time::Union{Int64, Missing}=missing, run_name::Union{String, Missing}=missing)
updaterun(instance::MLFlow, run::Run; status::Union{RunStatus, Missing}=missing,
    end_time::Union{Int64, Missing}=missing, run_name::Union{String, Missing}=missing)

Update Run metadata.

Arguments

instance: MLFlow configuration.

run_id: ID of the Run to update.

status: Updated status of the Run.

end_time: Unix timestamp in milliseconds of when the Run ended.

run_name: Updated name of the Run.

Returns

An instance of type RunInfo with the updated metadata.