MLFlowClient

MLFlowClient.MLFlowClient — Module

MLFlowClient

MLFlowClient is a Julia package for working with MLFlow using the REST API v2.0.

MLFlowClient allows you to create and manage MLFlow experiments, runs, and log metrics and artifacts. If you are not familiar with MLFlow and its concepts, please refer to MLFlow documentation.

Limitations

no authentication support.

when storing artifacts, the assumption is that MLFlow and this library run on the same server. Artifacts are stored using plain filesystem operations. Therefore, /mlruns or the specified artifact_location must be accessible to both the MLFlow server (read), and this library (write).

Installation

To install MLFlowClient, start a julia session and run the following:

julia> import Pkg

julia> Pkg.add("MLFlowClient")

or simply:

]add MLFlowClient