MLFlowClient

MLFlowClient.MLFlowClient — Module

MLFlowClient

MLFlowClient is a Julia package for working with MLFlow using the REST API v2.0.

MLFlowClient allows you to create and manage MLFlow experiments, runs, and log metrics and artifacts. If you are not familiar with MLFlow and its concepts, please refer to MLFlow documentation.

Installation

To install MLFlowClient, start a julia session and run the following:

julia> import Pkg

julia> Pkg.add("MLFlowClient")

or simply:

]add MLFlowClient