Reference

Types

MLFlowClient.MLFlow — Type

MLFlow(baseuri; apiversion)

Base type which defines location and version for MLFlow API service.

Fields

baseuri::String: base MLFlow tracking URI, e.g. http://localhost

apiversion: used API version, e.g. 2.0

Examples

julia> mlf = MLFlow("http://localhost")
MLFlow("http://localhost", 2.0)

MLFlowClient.MLFlowExperiment — Type

MLFlowExperiment

Represents an MLFlow experiment.

Fields

name::String: experiment name.

lifecycle_stage::String: life cycle stage, one of ["active", "deleted"]

experiment_id::Integer: experiment identifier.

tags::Any: list of tags.

artifact_location::String: where are experiment artifacts stored.

MLFlowClient.MLFlowRun — Type

MLFlowRun

Represents an MLFlow run.

Fields

info::MLFlowRunInfo: Run metadata.

data::MLFlowRunData: Run data.

MLFlowClient.MLFlowRunInfo — Type

MLFlowRunInfo

Represents run metadata.

Fields

run_id::String

experiment_id::Integer

status::MLFlowRunStatus

start_time::Union{Int64,Missing}

end_time::Union{Int64,Missing}

artifact_uri::String

lifecycle_stage::String

MLFlowClient.MLFlowRunData — Type

MLFlowRunData

Represents run data.

Fields

metrics

params

tags

TODO

Incomplete functionality.

MLFlowClient.MLFlowRunStatus — Type

MLFlowRunStatus

Represents the status of an MLFlow Run.

Fields

status::String: one of RUNNING/SCHEDULED/FINISHED/FAILED/KILLED

Experiments

MLFlowClient.createexperiment — Function

createexperiment(mlf::MLFlow; name=missing, artifact_location=missing, tags=missing)

Creates an MLFlow experiment.

Arguments

mlf: MLFlow configuration.

name: experiment name. If not specified, MLFlow sets it.

artifact_location: directory where artifacts of this experiment will be stored. If not specified, MLFlow uses its default configuration.

tags: a Dictionary of key-values which tag the experiment.

Returns

Experiment identifier (integer).

MLFlowClient.getexperiment — Function

getexperiment(mlf::MLFlow, experiment_id::Integer)

Retrieves an MLFlow experiment by id.

Arguments

mlf: MLFlow configuration.

experiment_id: Experiment identifier.

Returns

An instance of type MLFlowExperiment

getexperiment(mlf::MLFlow, experiment_name::String)

Retrieves an MLFlow experiment by name.

Arguments

mlf: MLFlow configuration.

experiment_name: Experiment name.

Returns

An instance of type MLFlowExperiment

MLFlowClient.listexperiments — Function

listexperiments(mlf::MLFlow)

Returns a list of MLFlow experiments.

TODO: not yet entirely implemented

MLFlowClient.deleteexperiment — Function

deleteexperiment(mlf::MLFlow, experiment_id)

Deletes an MLFlow experiment.

Arguments

mlf: MLFlow configuration.

experiment_id: experiment identifier.

Runs

MLFlowClient.createrun — Function

createrun(mlf::MLFlow, experiment_id; start_time=missing, tags=missing)

Creates a run associated to an experiment.

Arguments

mlf: MLFlow configuration.

experiment_id: experiment identifier.

Keywords

start_time: if provided, must be a UNIX timestamp in milliseconds. By default, set to current time.

tags: if provided, must be a key-value structure such as a dictionary.

Returns

an instance of type MLFlowRun

MLFlowClient.getrun — Function

getrun(mlf::MLFlow, run_id)

Retrieves information about an MLFlow run.

Arguments

mlf: MLFlow configuration.

run_id::String: run identifier.

Returns

an instance of type MLFlowRun

MLFlowClient.updaterun — Function

updaterun(mlf::MLFlow, run, status; end_time=missing)

Updates the status of an experiment's run.

Arguments

mlf: MLFlow configuration.

run: one of MLFlowRun, MLFlowRunInfo, or String.

status: either String and one of ["RUNNING", "SCHEDULED", "FINISHED", "FAILED", "KILLED"], or an instance of MLFlowRunStatus

Keywords

end_time: if provided, must be a UNIX timestamp in milliseconds. By default, set to current time.

MLFlowClient.deleterun — Function

deleterun(mlf::MLFlow, run)

Deletes an experiment's run.

Arguments

mlf: MLFlow configuration.

run: one of MLFlowRun, MLFlowRunInfo, or String.

Logging

MLFlowClient.logparam — Function

logparam(mlf::MLFlow, run, key, value)
logparam(mlf::MLFlow, run, kv)

Associates a key/value pair of parameters to the particular run.

Arguments

mlf: MLFlow configuration.

run: one of MLFlowRun, MLFlowRunInfo, or String.

key: parameter key (name).

value: parameter value.

One could also specify kv::Dict instead of separate key and value arguments.

MLFlowClient.logmetric — Function

logmetric(mlf::MLFlow, run, key, value::T; timestamp, step) where T<:Real
logmetric(mlf::MLFlow, run, key, values::AbstractArray{T}; timestamp, step) where T<:Real

Logs a metric value (or values) against a particular run.

Arguments

mlf: MLFlow configuration.

run: one of MLFlowRun, MLFlowRunInfo, or String

key: metric name.

value: metric value, must be numeric.

Keywords

timestamp: if provided, must be a UNIX timestamp in milliseconds. By default, set to current time.

step: step at which the metric value has been taken.

MLFlowClient.logartifact — Function

logartifact(mlf::MLFlow, run, filename)

Stores an artifact (file) in the run's artifact location.

Note

Assumes that artifact_uri is mapped to a local directory. At the moment, this only works if both MLFlow and the client are running on the same host or they map a directory that leads to the same location over NFS, for example.

Arguments

mlf::MLFlow: MLFlow onfiguration. Currently not used, but when this method is extended to support S3, information from mlf will be needed.

run: one of MLFlowRun, MLFlowRunInfo or String.

filename: path to the artifact that needs to be sent to MLFlow.