Types

MLFlowClient.MLFlow — Type

MLFlow

Base type which defines location and version for MLFlow API service.

Fields

apiroot::String: API root URL, e.g. http://localhost:5000/api

apiversion::Union{Integer, AbstractFloat}: used API version, e.g. 2.0

headers::Dict: HTTP headers to be provided with the REST API requests.

username::Union{Nothing, String}: username for basic authentication.

password::Union{Nothing, String}: password for basic authentication.

Warning

You cannot provide an Authorization header when an username and password are provided. An error will be thrown in that case.

Note

If MLFLOW_TRACKING_URI is set, the provided apiroot will be ignored.

If MLFLOW_TRACKING_USERNAME is set, the provided username will be ignored.

If MLFLOW_TRACKING_PASSWORD is set, the provided password will be ignored.

These indications will be displayed as warnings.

Examples

mlf = MLFlow()

remote_url="https://<your-server>.cloud.databricks.com"; # address of your remote server
mlf = MLFlow(remote_url, headers=Dict("Authorization" => "Bearer <your-secret-token>"))

MLFlowClient.Tag — Type

Tag <: LoggingData

Generic tag type for MLFlow entities.

Fields

key::String: The tag key.

value::String: The tag value.

MLFlowClient.Dataset — Type

Dataset

Represents a reference to data used for training, testing, or evaluation during the model development process.

Fields

name::String: The name of the dataset.

digest::String: The digest of the dataset.

source_type::String: The type of the dataset source.

source::String: Source information for the dataset.

schema::String: The schema of the dataset. This field is optional.

profile::String: The profile of the dataset. This field is optional.

MLFlowClient.DatasetInput — Type

DatasetInput

Represents a dataset and input tags.

Fields

tags::Array{Tag}: A list of tags for the dataset input.

dataset::Dataset: The dataset being used as a run input.

MLFlowClient.FileInfo — Type

FileInfo

Fields

path::String: Path relative to the root artifact directory run.

is_dir::Bool: Whether the path is a directory.

file_size::Int64: Size in bytes. Unset for directories.

MLFlowClient.ModelVersion — Type

ModelVersion

Fields

name::String: Unique name of the model.

version::String: Model’s version number.

creation_timestamp::Int64: Timestamp recorded when this model_version was created.

last_updated_timestamp::Int64: Timestamp recorded when metadata for this model_version was last updated.

user_id::Union{String, Nothing}: User that created this model_version.

current_stage::String: Current stage for this model_version.

description::String: Description of this model_version.

source::String: URI indicating the location of the source model artifacts, used when creating model_version.

run_id::String: MLflow run ID used when creating model_version, if source was generated by an experiment run stored in MLflow tracking server.

status::ModelVersionStatusEnum: Current status of model_version.

status_message::String: Details on current status, if it is pending or failed.

tags::Array{Tag}: Additional metadata key-value pairs.

run_link::String: Direct link to the run that generated this version. This field is set at model version creation time only for model versions whose source run is from a tracking server that is different from the registry server.

aliases::Array{String}: Aliases pointing to this model_version.

model_id::String: Optional model_id for ModelVersion that is used to link the RegisteredModel to the source logged model.

model_params::Array{ModelParam}: Optional parameters for the model.

model_metrics::Array{ModelMetric}: Optional metrics for the model.

deployment_job_state::ModelVersionDeploymentJobState: Deployment job state for this ModelVersion.

MLFlowClient.RegisteredModel — Type

RegisteredModel

Fields

name::String: Unique name for the model.

creation_timestamp::Int64: Timestamp recorded when this RegisteredModel was created.

last_updated_timestamp::Int64: Timestamp recorded when metadata for this RegisteredModel was last updated.

user_id::Union{String, Nothing}: User that created this RegisteredModel.

description::Union{String, Nothing}: Description of this RegisteredModel.

latest_versions::Array{ModelVersion}: Collection of latest model versions for each stage. Only contains models with current READY status.

tags::Array{Tag}: Additional metadata key-value pairs.

aliases::Array{RegisteredModelAlias}: Aliases pointing to model versions associated with this RegisteredModel.

deployment_job_id::String: Deployment job id for this model.

deployment_job_state::State: Deployment job state for this model.

MLFlowClient.RegisteredModelAlias — Type

RegisteredModelAlias

Alias for a registered model.

Fields

alias::String: The name of the alias.

version::String: The model version number that the alias points to.

MLFlowClient.Experiment — Type

Experiment

Fields

experiment_id::Integer: Unique identifier for the experiment.

name::String: Human readable name that identifies the experiment.

artifact_location::String: Location where artifacts for the experiment are stored.

lifecycle_stage::String: Current life cycle stage of the experiment: “active” or “deleted”. Deleted experiments are not returned by APIs.

last_update_time::Int64: Last update time.

creation_time::Int64: Creation time.

tags::Array{Tag}: Additional metadata key-value pairs.

MLFlowClient.Run — Type

Run

A single run.

Fields

info::RunInfo: Metadata of the run.

data::RunData: Run data (metrics, params, and tags).

inputs::RunInputs: Run inputs.

outputs::RunOutputs: Run outputs.

MLFlowClient.Param — Type

Param <: LoggingData

Param associated with a run.

Fields

key::String: Key identifying this param.

value::String: Value associated with this param.

MLFlowClient.Metric — Type

Metric <: LoggingData

Metric associated with a run, represented as a key-value pair.

Fields

key::String: Key identifying this metric.

value::Float64: Value associated with this metric.

timestamp::Int64: The timestamp at which this metric was recorded.

step::Union{Int64, Nothing}: Step at which to log the metric.

MLFlowClient.RunData — Type

RunInputs

Run data (metrics, params, and tags).

Fields

metrics::Array{Metric}: Run metrics.

params::Array{Param}: Run parameters.

tags::Array{Tag}: Additional metadata key-value pairs.

MLFlowClient.RunInfo — Type

RunInfo

Metadata of a single run.

Fields

run_id::String: Unique identifier for the run.

run_name::String: The name of the run.

experiment_id::String: The experiment ID.

status::RunStatus: Current status of the run.

start_time::Int64: Unix timestamp of when the run started in milliseconds.

end_time::Int64: Unix timestamp of when the run ended in milliseconds.

artifact_uri::String: URI of the directory where artifacts should be uploaded. This can be a local path (starting with “/”), or a distributed file system (DFS) path, like s3://bucket/directory or dbfs:/my/directory. If not set, the local ./mlruns directory is chosen.

lifecycle_stage::String: Current life cycle stage of the experiment: "active" or "deleted".

MLFlowClient.RunInputs — Type

RunInputs

Run inputs.

Fields

dataset_inputs::Array{DatasetInput}: Dataset inputs to the Run.

MLFlowClient.User — Type

User

Fields

id::String: User ID.

username::String: Username.

is_admin::Bool: Whether the user is an admin.

experiment_permissions::Array{ExperimentPermission}: All experiment permissions explicitly granted to the user.

registered_model_permissions::Array{RegisteredModelPermission}: All registered model explicitly granted to the user.

MLFlowClient.ExperimentPermission — Type

ExperimentPermission

Fields

experiment_id::String: Experiment id.

user_id::String: User id.

permission::Permission.PermissionEnum: Permission granted.

MLFlowClient.RegisteredModelPermission — Type

RegisteredModelPermission

Fields

name::String: RegisteredModel name.

user_id::String: User id.

permission::Permission.PermissionEnum: Permission granted.